delete ticket

+++ app/src/main/java/com/example/tubes_uts_e_2/api/ApiInterface.java
@@ -35,6 +35,6 @@ public interface ApiInterface {
 
     @Headers({"Accept: application/json"})
     @DELETE("ticket/{id}")
-    Call<TicketResponse> deleteTicket(@Path("id") long id);
+    Call<TicketResponse> deleteTicket(@Path("id") int id);
 
 }
